6433 | 9 files | M DumpSource2/schemas/client/CDOTA_Modifier_Visage_SummonFamiliars_In_Formation.h,  M DumpSource2/schemas/client/CQuickBuyController.h,  M DumpSource2/schemas/client/modifierfunction.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Visage_SummonFamiliars_In_Formation.h,  M DumpSource2/schemas/server/CQuickBuyController.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/18617073/

+++ DumpSource2/schemas/server/CDOTA_Modifier_Visage_SummonFamiliars_In_Formation.h
@@ -12,5 +12,5 @@ class CDOTA_Modifier_Visage_SummonFamiliars_In_Formation : public CDOTA_Buff
 	float32 recall_duration;
 	bool m_bTeleporting;
 	int32 familiar_attack_range;
-	int32 additional_target_search_radius;
+	float32 additional_target_search_radius;
 };
